Move MetadataTest to own class
